+++ docs/src/models/summarization/bm25.rst
@@ -1,7 +1,7 @@
-:mod:`summarization.bm25` -- BM25 ranking function
+:mod:`models.summarization.bm25` -- BM25 ranking function
 =========================================================
 
-.. automodule:: gensim.summarization.bm25
+.. automodule:: gensim.models.summarization.bm25
     :synopsis: BM25 ranking function
     :members:
     :inherited-members:
