+++ docs/src/models/summarization/commons.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.commons` -- Common graph functions
+=============================================================
+
+.. automodule:: gensim.models.summarization.commons
+    :synopsis: Common graph functions
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
