+++ docs/src/models/summarization/graph.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.graph` -- TextRank graph
+===================================================
+
+.. automodule:: gensim.models.summarization.graph
+    :synopsis: TextRank graph
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
