+++ docs/src/models/summarization/keywords.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.keywords` -- Keywords for TextRank summarization algorithm
+=====================================================================================
+
+.. automodule:: gensim.models.summarization.keywords
+    :synopsis: Keywords for TextRank summarization algorithm
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
