+++ docs/src/models/summarization/pagerank_weighted.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.pagerank_weighted` -- Weighted PageRank algorithm
+============================================================================
+
+.. automodule:: gensim.models.summarization.pagerank_weighted
+    :synopsis: Weighted PageRank algorithm
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
