+++ docs/src/models/summarization/summariser.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.summarizer` -- TextRank Summariser
+=============================================================
+
+.. automodule:: gensim.models.summarization.summarizer
+    :synopsis: TextRank Summariser
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
