+++ docs/src/models/summarization/syntactic_unit.rst
@@ -1,0 +1,9 @@
+:mod:`models.summarization.syntactic_unit` -- Syntactic Unit class
+==================================================================
+
+.. automodule:: gensim.models.summarization.syntactic_unit
+    :synopsis: Syntactic Unit class
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
