+++ docs/src/models/summarization/textcleaner.rst
@@ -1,0 +1,10 @@
+:mod:`models.summarization.textcleaner` -- Summarization pre-processing
+=======================================================================
+
+.. automodule:: gensim.models.summarization.textcleaner
+    :synopsis: Summarization pre-processing
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
+
