--- docs/src/summarization/commons.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.commons` -- Common graph functions
-=========================================================
-
-.. automodule:: gensim.summarization.commons
-    :synopsis: Common graph functions
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
