--- docs/src/summarization/graph.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.graph` -- TextRank graph
-=========================================================
-
-.. automodule:: gensim.summarization.graph
-    :synopsis: TextRank graph
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
