--- docs/src/summarization/keywords.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.keywords` -- Keywords for TextRank summarization algorithm
-==============================================================================
-
-.. automodule:: gensim.summarization.keywords
-    :synopsis: Keywords for TextRank summarization algorithm
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
