--- docs/src/summarization/pagerank_weighted.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.pagerank_weighted` -- Weighted PageRank algorithm
-=====================================================================
-
-.. automodule:: gensim.summarization.pagerank_weighted
-    :synopsis: Weighted PageRank algorithm
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
