--- docs/src/summarization/summariser.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.summarizer` -- TextRank Summariser
-=========================================================
-
-.. automodule:: gensim.summarization.summarizer
-    :synopsis: TextRank Summariser
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
