--- docs/src/summarization/syntactic_unit.rst
@@ -1,9 +1,0 @@
-:mod:`summarization.syntactic_unit` -- Syntactic Unit class
-===========================================================
-
-.. automodule:: gensim.summarization.syntactic_unit
-    :synopsis: Syntactic Unit class
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
