--- docs/src/summarization/textcleaner.rst
@@ -1,10 +1,0 @@
-:mod:`summarization.textcleaner` -- Summarization pre-processing
-================================================================
-
-.. automodule:: gensim.summarization.textcleaner
-    :synopsis: Summarization pre-processing
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
-
