+++ gensim/models/__init__.py
@@ -19,6 +19,7 @@ from .phrases import Phrases  # noqa:F401
 from .normmodel import NormModel  # noqa:F401
 from .atmodel import AuthorTopicModel  # noqa:F401
 from .ldaseqmodel import LdaSeqModel  # noqa:F401
+from .summarization import keywords, summarize, summarize_corpus  # noqa:F401
 
 from . import wrappers  # noqa:F401
